# Conflicts:
#	app/assets/javascripts/airesis.coffee
#	app/assets/javascripts/democracy.js
#	app/assets/stylesheets/application.scss
#	app/controllers/interest_borders_controller.rb
#	app/controllers/quorums_controller.rb
#	app/models/user.rb
#	app/views/quorums/new.html.erb

+++ app/models/alert.rb
@@ -123,6 +123,6 @@ class Alert < ActiveRecord::Base
   end
 
   def complete_alert_job
-    alert_job.destroy
+    alert_job.destroy if alert_job
   end
 end
